chore(#51): isolate Counter component
