+++ components/slides/domstol/NumberSlide/NumberSlide.tsx
@@ -1,31 +1,12 @@
-import { animate, motion } from 'framer-motion';
-import React, { useEffect, useRef } from 'react';
+import { motion } from 'framer-motion';
+import React from 'react';
 import { useInView } from 'react-intersection-observer';
+import Counter from '../../../common/Counter';
 import styles from './Number.module.scss';
 
 
 const NumberSlide = () => {
     const [inViewRef, inView] = useInView();
-
-    function Counter({ from, to }: { from: number, to: number }) {
-        const nodeRef = useRef() as React.MutableRefObject<HTMLParagraphElement>;
-
-        useEffect(() => {
-            if (inView) {
-                var node: any = nodeRef.current;
-
-                const controls = animate(from, to, {
-                    duration: 2,
-                    onUpdate(value) {
-                        node.textContent = Number(value.toFixed()).toLocaleString('no');
-                    }
-                });
-                return () => controls.stop();
-            }
-
-        }, [from, to]);
-        return <p className={styles.number} ref={nodeRef} />;
-    }
 
     return (
         <motion.section className={styles.section}
@@ -35,7 +16,7 @@ const NumberSlide = () => {
             transition={{ duration: 0.5 }}  >
             <div className={styles.extraheight}></div>
             <div className={styles.extraheight}></div>
-            <motion.p ref={inViewRef} className={styles.number}
+            <motion.div ref={inViewRef}
                 initial={{
                     color: '#ffffff',
                 }}
@@ -47,8 +28,8 @@ const NumberSlide = () => {
                     color: { duration: 2, delay: 2 },
                 }}
             >
-                <Counter from={0} to={261974} />
-            </motion.p>
+                <Counter from={0} to={261974} shouldAnimate={inView} className={styles.number} />
+            </motion.div>
             <div className={styles.extraheight}></div>
             <p className={styles.text1}>Straffede personer i Norge i 2020.</p>
             <div className={styles.extraheight}></div>
